@@ -1,6 +1,5 @@
-class Item < ActiveRecord::Base
-
-  acts_as_paranoid
+class Item < ApplicationRecord
+  # acts_as_paranoid
 
   belongs_to :article
 
