@@ -1,6 +1,6 @@
-class OrderTag < ActiveRecord::Base
+class OrderTag < ApplicationRecord
 
-  acts_as_paranoid
+  # acts_as_paranoid
 
   belongs_to :order
   belongs_to :tag
